Split WPT harness into files
